[14.0] [REF] Refactor account_reconciliation_widget files
