+++ account_reconciliation_widget/models/__init__.py
@@ -1,5 +1,6 @@
-from . import account_move
+from . import account_move_line
 from . import account_bank_statement
+from . import account_bank_statement_line
 from . import account_journal
 from . import reconciliation_widget
 from . import res_company
