--- account_reconciliation_widget/views/account_bank_statement_view.xml
@@ -1,25 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<odoo>
-    <data>
-        <record id="view_bank_statement_form" model="ir.ui.view">
-            <field name="name">account.bank.statement.inherit.view.form</field>
-            <field name="model">account.bank.statement</field>
-            <field name="inherit_id" ref="account.view_bank_statement_form" />
-            <field name="arch" type="xml">
-                <xpath expr="//button[@name='button_post']" position="before">
-                    <button
-                        name="action_bank_reconcile_bank_statements"
-                        string="Reconcile"
-                        type="object"
-                        class="oe_highlight"
-                        attrs="{'invisible':['|','|',('all_lines_reconciled','=',True),('line_ids','=',[]),('state', '!=', 'posted')]}"
-                    />
-                </xpath>
-                <xpath expr="//field[@name='date']" position="after">
-                    <field name="accounting_date" groups="base.group_no_one" />
-                </xpath>
-            </field>
-        </record>
-
-    </data>
-</odoo>
