+++ account_reconciliation_widget/views/account_journal.xml
